+++ packages/card/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0](https://github.com/inavac182/uireact/compare/@uireact/card@2.1.4...@uireact/card@3.0.0) (2024-04-27)
+
+**BREAKING CHANGE:** Adding framer-motion support for @UiCard
+
+
+
 ## 2.1.4 (2024-03-18)
 
 **Note:** Version bump only for package @uireact/card